shellb: use LS_COLORS to choose colors (if possible)
In future we may want to allow users to configure it from shellbrc, but
for now, when LS_COLORS is not availble, fallback to a sane default.
There is known issue with e.g. byobu, that forces different color scheme
that is not reflected in LS_COLORS envvar. Maybe we should query
dircolors?